backend/requirements.txt

Flask==2.2.2
Flask-RESTful==0.3.9
python-dotenv==0.15.0
Flask-SQLAlchemy==3.0.2
Flask-Migrate==4.0.4
Flask-Script==2.0.5
psycopg2==2.8.4
marshmallow==3.10.0
Flask-Cors==3.0.10
Flask-Testing==0.8.1
nose2==0.10.0
